Create A level Non-Exam Assessment
This is the code for the NEA project I did at A level. This project relied heavily on graphics and is programmed using Visual Basic. 
For my project, I designed and created a prototype for a program that is designed to help year 6 students prepare for their SATs exams. It had an initial test designed to evaluate a student's ability, then it goes on to lessons, questions and games for topics in their year 6 maths syllabus